Re-organize, move functionality into API
* wilfred.errors is now wilfred.api.errors, wilfred.database is now wilfred.api.database
