+++ wilfred/api/errors.py
